Make HW specs resistant to outdating (#20)
It should be sufficient to indicate the HW laptop brand + model as the #of cores and RAM size evolves over time. Similar to headsets, where Anker SoundCore Space is an outdated model, which has been replaced on the market. And some minor polishing has been done as a part of the PR.